basic services up and running. Learned something about definitions, templates and services. All services are starting including gemstone but somehow seaside support is missing in the extent

+++ vagrant/cookbooks/stone-creator/recipes/default.rb
@@ -13,27 +13,33 @@ include_recipe "gemstone"
 #    not_if { FileTest.exists?("/opt/stone-creator") }
 #end
 
-git "/opt/stone-creator" do
-   repository "https://git.gitorious.org/gemstone-tools/stone-creator.git"
-   reference "master"
-   action :sync
-end
+unless File.exists?("/opt/stone-creator")
+   git "/opt/stone-creator" do
+      repository "https://git.gitorious.org/gemstone-tools/stone-creator.git"
+      reference "master"
+      action :sync
+   end
 
-execute "creating application directory" do
-   command "mkdir /opt/application"
-   not_if { FileTest.exists?("/opt/application") }
-end
+   execute "creating application directory" do
+      command "mkdir /opt/application"
+      not_if { FileTest.exists?("/opt/application") }
+   end
 
-execute "creating default stone instance" do
-   command "/opt/stone-creator/bin/create-stone.sh -n stone-default -d /opt/application/stone-default"
-   not_if { FileTest.exists?("/opt/application/stone-default") }
-end
+   execute "creating default stone instance" do
+      command "/opt/stone-creator/bin/create-stone.sh -n stone-default -d /opt/application/stone-default -f"
+      not_if { FileTest.exists?("/opt/application/stone-default") }
+   end
 
-execute "creating symlink to /etc/init.d" do
-   command "ln -s /opt/application/stone-default/stone-default /etc/init.d/stone-default"
-   not_if { FileTest.exists?("/etc/init.d/stone-default") }
-end
+   execute "creating symlink to /etc/init.d" do
+      command "ln -s /opt/application/stone-default/stone-default /etc/init.d/stone-default"
+      not_if { FileTest.exists?("/etc/init.d/stone-default") }
+   end
 
-execute "activating stone-default" do
-   command "update-rc.d stone-default defaults"
+   execute "activating stone-default" do
+      command "update-rc.d stone-default defaults"
+   end
+
+   execute "starting stone-default" do
+      command "/etc/init.d/stone-default start"
+   end
 end
